replicator_tests.rs: refactor printing errors
This commit changes the construction of `FailureReason::RowNotMatching`, where we keep the index of the row that was mismatched. Previously we kept one-based index, but it may be confusing. Now we are keeping a zero-based index and converting it to one-based index in a function responsible for printing error messages. This change also allows us to be consistent with `FailureReason::TimestampsNotMatching`.