Update Scylla Rust Driver to 1.0
Most of the changes were trivial - different import paths, different
names of structs and methods, and some differences in errors.

The only non-obvious change is caused by ColumnType becoming
non_exhaustive. I had to handle "_" case in "update_or_insert" because
of that, for which I had to introduce an error type. This is in
scylla-cdc-replicator, so has no influence on public API.

While doing that I adeed Vector type handling in the same function.