add TabletsGenerationFetcher
add an implementation of GenerationFetcher for tablets-based keyspaces.

it has additional members keyspace_name and table_name for the table for
which we want to fetch generations and streams, because in tablets-based
keyspace each table has its own generations and streams, as opposed to
vnodes where they are global.

it fetches generation and streams from the system tables
system.cdc_timestamps and system.cdc_streams.